requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
peewee==3.17.8
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.16.3
pyTelegramBotAPI==4.24.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.10.0
urllib3==2.2.3
flake8==7.1.1
isort==5.13.2
mypy==1.13.0
wemake-python-styleguide==0.19.2
types-peewee==3.17.8.20241117